requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0

# Document processing
pypdf2>=3.0.1
python-docx>=0.8.11
python-pptx>=0.6.21

# Web interface
fastapi>=0.100.0
uvicorn>=0.22.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0